# packages/api-zod/src/generated/types/dataUnavailable.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DataUnavailableError } from './dataUnavailableError';

export interface DataUnavailable {
  error: DataUnavailableError;
}